added gravity & ground checking. Changed camera settings
it took a minute, but now gravity and the ground checking stuff is cool. I also changed the camera to be closer to how i want the final game camera angle to look. I'm going to have to figure out how to change camera angles and stuff, but that's a problem for future me.